@@ -1,6 +1,6 @@
 {
   "name": "lch.mercedesbenz.web",
-  "version": "0.0.3",
+  "version": "0.1.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.1.0",
+    "@angular/animations": "^17.3.12",
     "@angular/cdk": "^16.1.0",
-    "@angular/common": "^16.1.0",
-    "@angular/compiler": "^16.1.0",
-    "@angular/core": "^16.1.0",
-    "@angular/forms": "^16.1.0",
+    "@angular/common": "^17.3.12",
+    "@angular/compiler": "^17.3.12",
+    "@angular/core": "^17.3.12",
+    "@angular/forms": "^17.3.12",
     "@angular/material": "^16.1.0",
     "@angular/material-moment-adapter": "^16.1.0",
-    "@angular/platform-browser": "^16.1.0",
-    "@angular/platform-browser-dynamic": "^16.1.0",
-    "@angular/router": "^16.1.0",
+    "@angular/platform-browser": "^17.3.12",
+    "@angular/platform-browser-dynamic": "^17.3.12",
+    "@angular/router": "^17.3.12",
     "moment": "^2.29.4",
     "ngx-mat-select": "^16.0.4",
     "rxjs": "~7.8.0",
@@ -30,12 +30,12 @@
     "tslib": "^2.3.0",
     "uuid": "^9.0.0",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.3/xlsx-0.20.3.tgz",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.0",
-    "@angular/cli": "~16.1.0",
-    "@angular/compiler-cli": "^16.1.0",
+    "@angular-devkit/build-angular": "^17.3.12",
+    "@angular/cli": "~17.3.12",
+    "@angular/compiler-cli": "^17.3.12",
     "@types/jasmine": "~4.3.0",
     "@types/uuid": "^9.0.2",
     "autoprefixer": "^10.4.14",
@@ -47,6 +47,6 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "postcss": "^8.4.24",
     "tailwindcss": "^3.3.2",
-    "typescript": "~5.1.3"
+    "typescript": "~5.4.5"
   }
 }
