@@ -1,32 +1,42 @@
-import { Injectable } from "@angular/core";
-import { HttpClient } from "@angular/common/http";
-import { environment } from "src/environments/environment";
-import { map } from 'rxjs/operators';  
-import { PbiParameters } from "src/app/models/pbi-parameters/pbi-parameters.model";
-import { Observable } from "rxjs";
-@Injectable({ providedIn: 'root', }) 
-export class ReportService  {
-    HttpClient: any; 
-    private readonly controller = 'report';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { environment } from 'src/environments/environment';
+import { map } from 'rxjs/operators';
+import { PbiParameters } from 'src/app/models/pbi-parameters/pbi-parameters.model';
+import { Observable } from 'rxjs';
+import { BaseService } from '../base.service';
 
-    constructor(private http: HttpClient) { }
+@Injectable()
+export class ReportService extends BaseService<any> {
+  TAG = ReportService.name;
+  private readonly controller = 'report';
 
-    public GetToken(parameters: PbiParameters) {
-        const url = environment.api.reportUrl + 'PBI/GetToken';
-        return this.http.post(url, parameters).pipe(map((response: any) => {
-            return JSON.parse(response);
-        }));
-    }
+  constructor(httpClient: HttpClient) {
+    super(httpClient, 'report');
+  }
 
-    public getReport(
-        dateFrom: string | null,
-        dateTo: string,
-        pageNumber: number | undefined,
-        pageSize: number | undefined
-      ): Observable<any> {
-        const url: string = 
-        `${this.controller}/general?FechaPatentamientoDesde=${dateFrom ?? ''}&FechaPatentamientoHasta=${dateTo ?? ''}&PageSize=${pageSize}&PageNumber=${pageNumber}`;
-        return this.http.get(url);
-      }
-    
-} 
+  public GetToken(parameters: PbiParameters) {
+    const url = environment.api.reportUrl + 'PBI/GetToken';
+    return this.HttpClient.post(url, parameters).pipe(
+      map((response: any) => {
+        return JSON.parse(response);
+      })
+    );
+  }
+
+  public getReport(
+    dateFrom: string | null,
+    dateTo: string,
+    pageNumber: number | undefined,
+    pageSize: number | undefined
+  ): Observable<any> {
+    const url: string = `${
+      this.controller
+    }/commercial?FechaPatentamientoDesde=${
+      dateFrom ?? ''
+    }&FechaPatentamientoHasta=${
+      dateTo ?? ''
+    }&PageSize=${pageSize}&PageNumber=${pageNumber}`;
+    return this.HttpClient.get(url);
+  }
+}
