@@ -112,7 +112,9 @@ export class PermissionService extends BaseService<Permission> {
     );
   }
 
-  public getGranted(roleId: string | null): Observable<Permission[]> {
+  public getGranted(
+    roleId: string | null
+  ): Observable<Permission[] | undefined> {
     const user = this.authService.user;
     if (
       user &&
@@ -120,10 +122,12 @@ export class PermissionService extends BaseService<Permission> {
       user.role.permissions &&
       user.role.permissions.length > 0
     ) {
-      return new Observable((observer: Subscriber<Permission[]>) => {
-        observer.next(user?.role?.permissions);
-        observer.complete();
-      });
+      return new Observable(
+        (observer: Subscriber<Permission[] | undefined>) => {
+          observer.next(user?.role?.permissions);
+          observer.complete();
+        }
+      );
     }
     let url: string = `${this.controller}/granted?roleId=${roleId}`;
     return this.HttpClient.get<BaseResponse<Permission>>(url).pipe(
